refactor: add defaultNameOverride and reorganize
